Convert Model Objects between Forms

These are utility functions for model objects in the package MARSS-package. Users would not normally work directly with these functions.

Usage

marss_to_marxss(x, C.and.D.are.zero = FALSE)
marxss_to_marss(x, only.par = FALSE)

Arguments

x

An object of class marssMLE.

C.and.D.are.zero

If the C and D matrices are all 0, then a marss model can be converted to marxss without further information besides the marss model.

only.par

If only.par=TRUE then only the par element is changed and marss is used for the marss object.

Details

As the name of the functions imply, these convert marssMODEL objects of different forms into other forms. form=marss is the base form needed for the internal algorithms, thus other (more user friendly forms) must have a form_to_marss function to convert to the base form. The printing functions are customized to show output in the user-friendly form, thus a marss_to_form function is needed for print and coef methods for marssMLE objects.

Value

A marssMODEL object of the appropriate form.

Author

Eli Holmes, NOAA, Seattle, USA.

See also