Check inputs to MARSS call

This is a helper function to check the inputs to a MARSS() call for any errors. Not exported.

Usage

checkMARSSInputs( MARSS.inputs, silent = FALSE )

Arguments

MARSS.inputs

A list comprised of the needed inputs to a MARSS call: data, inits, model, control, method, form)

silent

Suppresses printing of progress bars, error messages, warnings and convergence information.

Details

This is a helper function to check that all the inputs to a MARSS() function call are properly specified.

If arguments inits or control are not provided by the user, they will be set by the alldefaults[[method]] object specified in MARSSsettings. Argument model specifies the model structure using a list of matrices; see MARSS or the User Guide for instructions on how to specify model structure. If model is left off, then the function MARSS.form() is used to determine the default model structure.

Value

If the function does not stop due to errors, it returns an updated list with elements

data

Data supplied by user.

model

Not changed. Will be updated by the MARSS.form function (e.g. MARSS.marxss).

inits

A list specifying initial values for parameters to be used at iteration 1 in iterative maximum-likelihood algorithms.

method

The method used for estimation.

form

The equation form used to convert wrapper object to a marssMODEL object.

control

See Arguments.

Author

Kellie Wills, NOAA, Seattle, USA.

See also