Return brief summary information on a MARSS fit

This returns a data frame with brief summary information.

coef.det

The coefficient of determination. This is the squared correlation between the fitted values and the original data points. This is simply a metric for the difference between the data points and the fitted values and should not be used for formal model comparison.

sigma

The sample variance (unbiased) of the data residuals (fitted minus data). This is another simple metric of the difference between the data and fitted values. This is different than the sigma returned by an arima() call for example. That sigma would be akin to sqrt(Q) in the MARSS parameters; 'akin' because MARSS models are multivariate and the sigma returned by arima() is for a univariate model.

df

The number of estimated parameters. Denoted num.params in a marssMLE object.

logLik

The log-likelihood.

AIC

Akaike information criterion.

AICc

Akaike information criterion corrected for small sample size.

AICbb

Non-parametric bootstrap Akaike information criterion if in the marssMLE object.

AICbp

Parametric bootstrap Akaike information criterion if in the marssMLE object.

convergence

0 if converged according to the convergence criteria set. Note the default convergence criteria are high in order to speed up fitting. A number other than 0 means the model did not meet the convergence criteria.

errors

0 if no errors. 1 if some type of error or warning returned.

Usage

# S3 method for marssMLE
glance(x, ...)

Arguments

x

A marssMLE object

...

Not used.

Examples

dat <- t(harborSeal)
dat <- dat[c(2, 11, 12), ]
fit <- MARSS(dat, model = list(Z = factor(c("WA", "OR", "OR"))))
#> Success! abstol and log-log tests passed at 37 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 37 iterations. 
#> Log-likelihood: 13.72233 
#> AIC: -11.44465   AICc: -8.918339   
#>  
#>                 Estimate
#> A.OR.SouthCoast  0.49280
#> R.diag           0.02509
#> U.WA             0.06171
#> U.OR             0.03686
#> Q.(WA,WA)        0.01082
#> Q.(OR,OR)        0.00439
#> x0.WA            7.41712
#> x0.OR            6.56460
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 

glance(fit)
#>    coef.det      sigma df   logLik       AIC      AICc convergence errors
#> 1 0.9011206 0.03770242  8 13.72233 -11.44465 -8.918339           0      0