Class "marssMLE"

marssMLE objects specify fitted multivariate autoregressive state-space models (maximum-likelihood) in the package MARSS-package.

A marssMLE object in the MARSS-package that has all the elements needed for maximum-likelihood estimation of multivariate autoregressive state-space model: the data, model, estimation methods, and any control options needed for the method. If the model has been fit and parameters estimated, the object will also have the MLE parameters. Other functions add other elements to the marssMLE object, such as CIs, s.e.'s, AICs, and the observed Fisher Information matrix. There are print, summary, coef, fitted, residuals, predict and simulate methods for marssMLE objects and a bootstrap function. Rather than working directly with the elements of a marssMLE object, use print(), tidy(), fitted(), tsSmooth(), predict(), or residuals() to extract output.

Methods

print

signature(x = "marssMLE"): ...

summary

signature(object = "marssMLE"): ...

coef

signature(object = "marssMLE"): ...

residuals

signature(object = "marssMLE"): ...

predict

signature(object = "marssMLE"): ...

fitted

signature(object = "marssMLE"): ...

logLik

signature(object = "marssMLE"): ...

simulate

signature(object = "marssMLE"): ...

forecast

signature(object = "marssMLE"): ...

accuracy

signature(object = "marssMLE"): ...

toLatex

signature(object = "marssMLE"): ...

See also

Author

Eli Holmes and Kellie Wills, NOAA, Seattle, USA