Return estimated parameters with summary information

tidy.marssMLE is the method for the tidy generic. It returns the parameter estimates and their confidence intervals.

Usage

# S3 method for marssMLE
tidy(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x

a marssMLE object

conf.int

Whether to compute confidence and prediction intervals on the estimates.

conf.level

Confidence level. alpha=1-conf.level

...

Optional arguments. If conf.int=TRUE, then arguments to specify how CIs are computed can be passed in. See details and MARSSparamCIs.

Value

A data frame with estimates, sample standard errors, and confidence intervals.

Details

tidy.marssMLE() assembles information available via the print() and coef() functions into a data frame that summarizes the estimates. If conf.int=TRUE, MARSSparamCIs() will be run to add confidence intervals to the model object if these are not already added. The default CIs are calculated using a analytically computed Hessian matrix. This can be changed by passing in optional arguments for MARSSparamCIs().

Examples

dat <- t(harborSeal)
dat <- dat[c(2, 11, 12), ]
fit <- MARSS(dat)
#> Success! abstol and log-log tests passed at 55 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 55 iterations. 
#> Log-likelihood: 30.86888 
#> AIC: -41.73776   AICc: -37.73776   
#>  
#>                                           Estimate
#> R.diag                                     0.00450
#> U.X.CoastalEstuaries                       0.06050
#> U.X.OR.NorthCoast                          0.05227
#> U.X.OR.SouthCoast                          0.02148
#> Q.(X.CoastalEstuaries,X.CoastalEstuaries)  0.02499
#> Q.(X.OR.NorthCoast,X.OR.NorthCoast)        0.01994
#> Q.(X.OR.SouthCoast,X.OR.SouthCoast)        0.00297
#> x0.X.CoastalEstuaries                      7.37247
#> x0.X.OR.NorthCoast                         6.26598
#> x0.X.OR.SouthCoast                         7.40658
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 

# A data frame of the estimated parameters
tidy(fit)
#>                                         term    estimate   std.error
#> 1                                     R.diag 0.004495377 0.002085693
#> 2                       U.X.CoastalEstuaries 0.060504720 0.032491337
#> 3                          U.X.OR.NorthCoast 0.052266571 0.027915656
#> 4                          U.X.OR.SouthCoast 0.021484041 0.011366026
#> 5  Q.(X.CoastalEstuaries,X.CoastalEstuaries) 0.024993374 0.009890095
#> 6        Q.(X.OR.NorthCoast,X.OR.NorthCoast) 0.019941851 0.008221337
#> 7        Q.(X.OR.SouthCoast,X.OR.SouthCoast) 0.002965429 0.001994353
#> 8                      x0.X.CoastalEstuaries 7.372465940 0.174049400
#> 9                         x0.X.OR.NorthCoast 6.265982529 0.268689177
#> 10                        x0.X.OR.SouthCoast 7.406580785 0.136833655
#>         conf.low     conf.up
#> 1   0.0004074930 0.008583260
#> 2  -0.0031771304 0.124186570
#> 3  -0.0024471090 0.106980250
#> 4  -0.0007929603 0.043761042
#> 5   0.0056091450 0.044377603
#> 6   0.0038283270 0.036055374
#> 7  -0.0009434297 0.006874289
#> 8   7.0313353843 7.713596496
#> 9   5.7393614197 6.792603639
#> 10  7.1383917498 7.674769821