MARSScv is a wrapper for MARSS that re-fits the model with cross validated data.

MARSScv is a wrapper for MARSS that re-fits the model with cross validated data.

Usage

MARSScv(
  y,
  model = NULL,
  inits = NULL,
  method = "kem",
  form = "marxss",
  silent = FALSE,
  control = NULL,
  fun.kf = c("MARSSkfas", "MARSSkfss"),
  fold_ids = NULL,
  future_cv = FALSE,
  n_future_cv = floor(ncol(y)/3),
  interval = "confidence",
  ...
)

Arguments

y

A n x T matrix of n time series over T time steps. Only y is required for the function. A ts object (univariate or multivariate) can be used and this will be converted to a matrix with time in the columns.

model

Model specification using a list of parameter matrix text shortcuts or matrices. See Details and MARSS.marxss() for the default form. Or better yet open the Quick Start Guide RShowDoc("Quick_Start",package="MARSS")

inits

A list with the same form as the list output by coef(fit) that specifies initial values for the parameters. See also MARSS.marxss().

method

Estimation method. MARSS provides an EM algorithm (method="kem") (see MARSSkem()) and the BFGS algorithm (method="BFGS") (see MARSSoptim()). Fast TMB fitting provided by the companion package marssTMB.

form

The equation form used in the MARSS() call. The default is "marxss". See MARSS.marxss() or MARSS.dfa()

silent

Setting to TRUE(1) suppresses printing of full error messages, warnings, progress bars and convergence information. Setting to FALSE(0) produces error output. Setting silent=2 will produce more verbose error messages and progress information.

control

Estimation options for the maximization algorithm. The typically used control options for method="kem" are below but see marssMLE for the full list of control options. Note many of these are not allowed if method="BFGS"; see MARSSoptim() for the allowed control options for this method.

fun.kf

What Kalman filter function to use. MARSS has two: MARSSkfas() which is based on the Kalman filter in the KFAS package based on Koopman and Durbin and MARSSkfss() which is a native R implementation of the Kalman filter and smoother in Shumway and Stoffer. The KFAS filter is much faster. MARSSkfas() modifies the input and output in order to output the lag-one covariance smoother needed for the EM algorithm (per page 321 in Shumway and Stoffer (2000).

fold_ids

A n x T matrix of integers, with values assigned by the user to folds. If not included, data are randomly assigned to one of 10 folds

future_cv

Whether or not to use future cross validation (defaults to FALSE), where data up to time T-1 are used to predict data at time T. Data are held out by time slices, and the fold_ids argument is ignored.

n_future_cv

Number of time slices to hold out for future cross validation. Defaults to floor(n_future_cv/3). Predictions are made for the last n_future_cv time steps

interval

uncertainty interval for prediction. Can be one of "confidence" or "prediction", and defaults to "confidence"

...

not used

Value

A list object, containing cv_pred (a matrix of predictions), cv_se (a matrix of SEs), fold_ids (a matrix of fold ids used as data), and df (a dataframe containing the original data, predictions, SEs, and folds)

Examples

# \donttest{
dat <- t(harborSealWA)
dat <- dat[2:4, ] # remove the year row
# fit a model with 1 hidden state and 3 observation time series
# cross validation here is random, 10 folds
fit <- MARSScv(dat, model = list(
  Z = matrix(1, 3, 1),
  R = "diagonal and equal"
))
#> Success! abstol and log-log tests passed at 22 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 22 iterations. 
#> Log-likelihood: 14.04175 
#> AIC: -16.0835   AICc: -14.13001   
#>  
#>         Estimate
#> A.SJI    0.81338
#> A.EBays  0.28790
#> R.diag   0.02339
#> U.U      0.06167
#> Q.Q      0.00825
#> x0.x0    6.10031
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 21 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 21 iterations. 
#> Log-likelihood: 15.6137 
#> AIC: -19.22741   AICc: -17.22741   
#>  
#>         Estimate
#> A.SJI    0.78342
#> A.EBays  0.27589
#> R.diag   0.02112
#> U.U      0.06176
#> Q.Q      0.00842
#> x0.x0    6.10823
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 28 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 28 iterations. 
#> Log-likelihood: 14.43699 
#> AIC: -16.87398   AICc: -14.60371   
#>  
#>         Estimate
#> A.SJI    0.83527
#> A.EBays  0.31649
#> R.diag   0.02003
#> U.U      0.05850
#> Q.Q      0.00858
#> x0.x0    6.12355
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 20 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 20 iterations. 
#> Log-likelihood: 17.4562 
#> AIC: -22.91239   AICc: -21.0033   
#>  
#>         Estimate
#> A.SJI     0.8159
#> A.EBays   0.2570
#> R.diag    0.0210
#> U.U       0.0561
#> Q.Q       0.0069
#> x0.x0     6.2143
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 21 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 21 iterations. 
#> Log-likelihood: 13.17189 
#> AIC: -14.34377   AICc: -12.39028   
#>  
#>         Estimate
#> A.SJI    0.79792
#> A.EBays  0.27246
#> R.diag   0.02383
#> U.U      0.06074
#> Q.Q      0.00926
#> x0.x0    6.10820
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 22 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 22 iterations. 
#> Log-likelihood: 14.42321 
#> AIC: -16.84641   AICc: -14.89292   
#>  
#>         Estimate
#> A.SJI    0.78654
#> A.EBays  0.25516
#> R.diag   0.02245
#> U.U      0.06077
#> Q.Q      0.00922
#> x0.x0    6.10960
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 23 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 23 iterations. 
#> Log-likelihood: 10.93771 
#> AIC: -9.875419   AICc: -7.721573   
#>  
#>         Estimate
#> A.SJI    0.80137
#> A.EBays  0.29445
#> R.diag   0.02485
#> U.U      0.06071
#> Q.Q      0.00922
#> x0.x0    6.10123
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 22 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 22 iterations. 
#> Log-likelihood: 13.754 
#> AIC: -15.508   AICc: -13.408   
#>  
#>         Estimate
#> A.SJI     0.7874
#> A.EBays   0.2410
#> R.diag    0.0206
#> U.U       0.0654
#> Q.Q       0.0116
#> x0.x0     6.0039
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 23 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 23 iterations. 
#> Log-likelihood: 11.11644 
#> AIC: -10.23288   AICc: -8.022352   
#>  
#>         Estimate
#> A.SJI    0.77718
#> A.EBays  0.27801
#> R.diag   0.02517
#> U.U      0.06351
#> Q.Q      0.00764
#> x0.x0    6.11946
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 23 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 23 iterations. 
#> Log-likelihood: 13.148 
#> AIC: -14.296   AICc: -12.08547   
#>  
#>         Estimate
#> A.SJI     0.7710
#> A.EBays   0.2835
#> R.diag    0.0197
#> U.U       0.0597
#> Q.Q       0.0131
#> x0.x0     6.0976
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 

# second, demonstrate passing in pre-specified folds
fold_ids <- matrix(
  sample(1:5, size = nrow(dat) * ncol(dat), replace = TRUE),
  nrow(dat), ncol(dat)
)
fit <- MARSScv(dat, model = list(
  Z = matrix(1, 3, 1),
  R = "diagonal and equal"
), fold_ids = fold_ids)
#> Success! abstol and log-log tests passed at 30 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 30 iterations. 
#> Log-likelihood: 11.55173 
#> AIC: -11.10345   AICc: -8.478452   
#>  
#>         Estimate
#> A.SJI     0.8272
#> A.EBays   0.2619
#> R.diag    0.0195
#> U.U       0.0611
#> Q.Q       0.0113
#> x0.x0     6.0744
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 28 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 28 iterations. 
#> Log-likelihood: 12.43474 
#> AIC: -12.86947   AICc: -10.24447   
#>  
#>         Estimate
#> A.SJI     0.7930
#> A.EBays   0.2575
#> R.diag    0.0191
#> U.U       0.0582
#> Q.Q       0.0101
#> x0.x0     6.1753
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 22 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 22 iterations. 
#> Log-likelihood: 12.65615 
#> AIC: -13.3123   AICc: -11.2123   
#>  
#>         Estimate
#> A.SJI    0.79513
#> A.EBays  0.26282
#> R.diag   0.02372
#> U.U      0.06106
#> Q.Q      0.00821
#> x0.x0    6.11202
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 27 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 27 iterations. 
#> Log-likelihood: 10.78428 
#> AIC: -9.568566   AICc: -7.097978   
#>  
#>         Estimate
#> A.SJI    0.80969
#> A.EBays  0.31874
#> R.diag   0.02271
#> U.U      0.06356
#> Q.Q      0.00919
#> x0.x0    6.08028
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 23 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 23 iterations. 
#> Log-likelihood: 13.46221 
#> AIC: -14.92442   AICc: -12.77058   
#>  
#>         Estimate
#> A.SJI    0.75374
#> A.EBays  0.27936
#> R.diag   0.02158
#> U.U      0.06074
#> Q.Q      0.00951
#> x0.x0    6.11469
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 

# third, illustrate future cross validation
fit <- MARSScv(dat, model = list(
  Z = matrix(1, 3, 1),
  R = "diagonal and equal"
), future_cv = TRUE, n_future_cv = 5)
#> Success! abstol and log-log tests passed at 21 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 21 iterations. 
#> Log-likelihood: 14.91458 
#> AIC: -17.82916   AICc: -15.87567   
#>  
#>         Estimate
#> A.SJI    0.80143
#> A.EBays  0.28885
#> R.diag   0.02367
#> U.U      0.06778
#> Q.Q      0.00652
#> x0.x0    6.10448
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 27 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 27 iterations. 
#> Log-likelihood: 16.33737 
#> AIC: -20.67474   AICc: -18.57474   
#>  
#>         Estimate
#> A.SJI    0.79451
#> A.EBays  0.30766
#> R.diag   0.02326
#> U.U      0.07594
#> Q.Q      0.00345
#> x0.x0    6.11259
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 32 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 32 iterations. 
#> Log-likelihood: 17.44618 
#> AIC: -22.89236   AICc: -20.62209   
#>  
#>         Estimate
#> A.SJI    0.80612
#> A.EBays  0.33869
#> R.diag   0.02147
#> U.U      0.08159
#> Q.Q      0.00269
#> x0.x0    6.09125
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 32 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 32 iterations. 
#> Log-likelihood: 15.81173 
#> AIC: -19.62347   AICc: -17.15288   
#>  
#>         Estimate
#> A.SJI    0.79297
#> A.EBays  0.34634
#> R.diag   0.02173
#> U.U      0.08298
#> Q.Q      0.00297
#> x0.x0    6.08537
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
#> Success! abstol and log-log tests passed at 29 iterations.
#> Alert: conv.test.slope.tol is 0.5.
#> Test with smaller values (<0.1) to ensure convergence.
#> 
#> MARSS fit is
#> Estimation method: kem 
#> Convergence test: conv.test.slope.tol = 0.5, abstol = 0.001
#> Estimation converged in 29 iterations. 
#> Log-likelihood: 16.2755 
#> AIC: -20.55101   AICc: -17.84133   
#>  
#>         Estimate
#> A.SJI    0.79445
#> A.EBays  0.38057
#> R.diag   0.01893
#> U.U      0.08406
#> Q.Q      0.00393
#> x0.x0    6.05363
#> Initial states (x0) defined at t=0
#> 
#> Standard errors have not been calculated. 
#> Use MARSSparamCIs to compute CIs and bias estimates.
#> 
# }